Moved containers status to separated file
