+++ core-ui/src/components/Predefined/Details/Pod/ContainersData.js
@@ -2,11 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import LuigiClient from '@luigi-project/client';
 import { useTranslation } from 'react-i18next';
-import { StatusBadge } from 'react-shared';
 
 import { LayoutPanel, Button } from 'fundamental-react';
 import './ContainersData.scss';
 import { LayoutPanelRow } from 'shared/components/LayoutPanelRow/LayoutPanelRow';
+import { ContainerStatus } from './ContainerStatus';
 
 const getPorts = ports => {
   if (ports?.length) {
@@ -20,47 +20,6 @@ const getPorts = ports => {
     return '';
   }
 };
-
-function GetContainerStatus({ status }) {
-  const { i18n } = useTranslation();
-
-  const state =
-    status.state.running || status.state.waiting || status.state.terminated;
-
-  const message = state.message || null;
-
-  const badgeType = status => {
-    switch (status) {
-      case 'Running':
-      case 'Completed':
-      case 'Succeeded':
-        return 'success';
-      case 'ContainerCreating':
-      case 'Initing':
-      case 'Pending':
-      case 'PodInitializing':
-      case 'Terminating':
-        return 'info';
-      case 'Unknown':
-        return undefined;
-      default:
-        return 'error';
-    }
-  };
-
-  return (
-    <div>
-      <StatusBadge
-        i18n={i18n}
-        resourceKind="containers"
-        additionalContent={message}
-        type={badgeType(state.reason || 'Running')}
-      >
-        {state.reason || 'Running'}
-      </StatusBadge>
-    </div>
-  );
-}
 
 ContainersData.propTypes = {
   containers: PropTypes.arrayOf(PropTypes.object).isRequired,
@@ -87,7 +46,7 @@ export default function ContainersData({ type, containers, statuses }) {
       <LayoutPanel.Body>
         <LayoutPanelRow
           name={t('common.headers.status')}
-          value={<GetContainerStatus status={status} />}
+          value={<ContainerStatus status={status} />}
         />
         {container.image && (
           <LayoutPanelRow
